Added TestClass
