+++ ros/src/computing/planning/decision/packages/decision_maker/include/decision_maker_node.hpp
@@ -112,7 +112,7 @@ struct AutowareStatus
 
 class DecisionMakerNode
 {
-  friend class TestSuite;
+  friend class TestClass;
 private:
   ros::NodeHandle nh_;
   ros::NodeHandle private_nh_;
